Cảm ơn AI

+++ EVService/src/main/java/edu/uth/evservice/EVService/requests/ServiceItemPartRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ServiceItemPartRequest {
-    private Long serviceItemId;
-    private Long partId;
+    private Integer serviceItemId;
+    private Integer partId;
     private Integer quantity;
 }
