+++ EVService/src/main/java/edu/uth/evservice/EVService/services/IPartService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface IPartService {
     List<PartDto> getAllParts();
-    PartDto getPartById(Long id);
+    PartDto getPartById(Integer id);
     PartDto createPart(PartRequest request);
-    PartDto updatePart(Long id, PartRequest request);
-    void deletePart(Long id);
+    PartDto updatePart(Integer id, PartRequest request);
+    void deletePart(Integer id);
 }
